pyproject.toml
[tool.poetry]
name = "social"
version = "0.1.0"
description = "poetry config"
authors = ["PL"]

[tool.poetry.dependencies]
python = "3.6.7"
django = "^2.2"
django_comments_xtd = "^2.4"
django_summernote = "^0.8.11"
django-allauth = "^0.40.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
